Fix crash on 32 bit architectures

+++ snapd-glib/snapd-user-information.h
@@ -24,7 +24,7 @@ G_BEGIN_DECLS
 
 G_DECLARE_FINAL_TYPE (SnapdUserInformation, snapd_user_information, SNAPD, USER_INFORMATION, GObject)
 
-int            snapd_user_information_get_id        (SnapdUserInformation *user_information);
+gint64         snapd_user_information_get_id        (SnapdUserInformation *user_information);
 
 const gchar   *snapd_user_information_get_username  (SnapdUserInformation *user_information);
